Replace float by int32.

+++ Source/Engine/Core/Math/Int4.cpp
@@ -15,35 +15,35 @@ const Int4 Int4::One(1);
 const Int4 Int4::Minimum(MIN_int32);
 const Int4 Int4::Maximum(MAX_int32);
 
-Int4::Int4(const Int2& xy, float z, float w)
+Int4::Int4(const Int2& xy, int32 z, int32 w)
     : X(xy.X)
     , Y(xy.Y)
-    , Z(static_cast<int32>(z))
-    , W(static_cast<int32>(w))
+    , Z(z)
+    , W(w)
 {
 }
 
-Int4::Int4(const Int3& xyz, float w)
+Int4::Int4(const Int3& xyz, int32 w)
     : X(xyz.X)
     , Y(xyz.Y)
     , Z(xyz.Z)
-    , W(static_cast<int32>(w))
+    , W(w)
 {
 }
 
-Int4::Int4(const Vector2& v, float z, float w)
+Int4::Int4(const Vector2& v, int32 z, int32 w)
     : X(static_cast<int32>(v.X))
     , Y(static_cast<int32>(v.Y))
-    , Z(static_cast<int32>(z))
-    , W(static_cast<int32>(w))
+    , Z(z)
+    , W(w)
 {
 }
 
-Int4::Int4(const Vector3& v, float w)
+Int4::Int4(const Vector3& v, int32 w)
     : X(static_cast<int32>(v.X))
     , Y(static_cast<int32>(v.Y))
     , Z(static_cast<int32>(v.Z))
-    , W(static_cast<int32>(w))
+    , W(w)
 {
 }
